# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          $ref: "#/components/responses/Error"
  /admin/users:
    get:
      operationId: getUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
  /admin/users/{id}/approve:
    post:
      operationId: approveUser
      tags: [admin]
      summary: Approve a pending user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/users/{id}/reject:
    post:
      operationId: rejectUser
      tags: [admin]
      summary: Reject a user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/users/{id}/password:
    post:
      operationId: setUserPassword
      tags: [admin]
      summary: Change a user's password
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordInput"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/keys:
    get:
      operationId: getKeys
      tags: [admin]
      summary: List all keys
      parameters:
        - name: plan
          in: query
          required: false
          schema:
            type: string
            enum: [trial, daily, weekly, monthly, lifetime]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [inactive, active, expired, revoked]
      responses:
        "200":
          description: Keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: generateKeys
      tags: [admin]
      summary: Generate one or many keys (optionally assigned to a user email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyGenerateInput"
      responses:
        "201":
          description: Generated keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Key"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /admin/keys/{id}/revoke:
    post:
      operationId: revokeKey
      tags: [admin]
      summary: Revoke a key
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/keys/{id}/reset-device:
    post:
      operationId: resetKeyDevice
      tags: [admin]
      summary: Reset the device binding of a key
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Updated key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/keys/{id}:
    delete:
      operationId: deleteKey
      tags: [admin]
      summary: Delete a key
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/managers:
    get:
      operationId: getManagers
      tags: [admin]
      summary: List manager accounts (admin only)
      responses:
        "200":
          description: Managers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: createManager
      tags: [admin]
      summary: Create a manager account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagerInput"
      responses:
        "201":
          description: Manager created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /admin/managers/{id}:
    delete:
      operationId: deleteManager
      tags: [admin]
      summary: Delete a manager account (admin only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/promotions:
    get:
      operationId: getPromotions
      tags: [admin]
      summary: List promotions (admin only)
      responses:
        "200":
          description: Promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
        "401":
          $ref: "#/components/responses/Error"
    post:
      operationId: createPromotion
      tags: [admin]
      summary: Create a promotion (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionInput"
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /admin/promotions/{id}:
    delete:
      operationId: deletePromotion
      tags: [admin]
      summary: Delete a promotion (admin only)
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /admin/plans/{plan}/price:
    put:
      operationId: setPlanPrice
      tags: [admin]
      summary: Change the base price of a plan (admin only)
      parameters:
        - name: plan
          in: path
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, lifetime]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanPriceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    PasswordInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 6
...
    ManagerInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    PromotionInput:
      type: object
      required: [plan, priceCents, durationHours]
      properties:
        plan:
          type: string
          enum: [daily, weekly, monthly, lifetime]
        priceCents:
          type: integer
          minimum: 50
        durationHours:
          type: integer
          minimum: 1
          maximum: 8760
        bannerText:
          type: ["string", "null"]
    Promotion:
      type: object
      required: [id, plan, priceCents, bannerText, endsAt, createdAt, active]
      properties:
        id:
          type: string
        plan:
          type: string
          enum: [daily, weekly, monthly, lifetime]
        priceCents:
          type: integer
        bannerText:
          type: ["string", "null"]
        endsAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        active:
          type: boolean
    PlanPriceInput:
      type: object
      required: [priceCents]
      properties:
        priceCents:
          type: integer
          minimum: 50
...
    KeyGenerateInput:
      type: object
      required: [plan, quantity]
      properties:
        plan:
          type: string
          enum: [trial, daily, weekly, monthly, lifetime]
        quantity:
          type: integer
          minimum: 1
          maximum: 500
        userEmail:
          type: ["string", "null"]
          format: email
...
    User:
      type: object
      required: [id, name, email, role, status, keyCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, client]
        status:
          type: string
          enum: [pending, approved, rejected]
        keyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    Key:
      type: object
      required: [id, code, plan, status, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        plan:
          type: string
          enum: [trial, daily, weekly, monthly, lifetime]
        status:
          type: string
          enum: [inactive, active, expired, revoked]
        userId:
          type: ["string", "null"]
        userEmail:
          type: ["string", "null"]
        deviceFingerprint:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        activatedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    AdminStats:
      type: object
      required:
        - totalUsers
        - pendingUsers
        - approvedUsers
        - totalKeys
        - activeKeys
        - inactiveKeys
        - expiredKeys
        - revokedKeys
        - planCounts
      properties:
        totalUsers:
          type: integer
        pendingUsers:
          type: integer
        approvedUsers:
          type: integer
        totalKeys:
          type: integer
        activeKeys:
          type: integer
        inactiveKeys:
          type: integer
        expiredKeys:
          type: integer
        revokedKeys:
          type: integer
        planCounts:
          $ref: "#/components/schemas/PlanCounts"
    PlanCounts:
      type: object
      required: [trial, daily, weekly, monthly, lifetime]
      properties:
        trial:
          type: integer
        daily:
          type: integer
        weekly:
          type: integer
        monthly:
          type: integer
        lifetime:
          type: integer
...