# lib/api-zod/src/generated/types/key.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KeyPlan } from './keyPlan';
import type { KeyStatus } from './keyStatus';

export interface Key {
  id: string;
  code: string;
  plan: KeyPlan;
  status: KeyStatus;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  deviceFingerprint?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  activatedAt?: Date | null;
  /** @nullable */
  expiresAt?: Date | null;
  createdAt: Date;
}