# lib/api-zod/src/generated/types/promotion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PromotionPlan } from './promotionPlan';

export interface Promotion {
  id: string;
  plan: PromotionPlan;
  priceCents: number;
  /** @nullable */
  bannerText: string | null;
  endsAt: Date;
  createdAt: Date;
  active: boolean;
}