# lib/api-zod/src/generated/types/promotionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PromotionInputPlan } from './promotionInputPlan';

export interface PromotionInput {
  plan: PromotionInputPlan;
  /** @minimum 50 */
  priceCents: number;
  /**
     * @minimum 1
     * @maximum 8760
     */
  durationHours: number;
  /** @nullable */
  bannerText?: string | null;
}